+++ .gitignore
@@ -9,3 +9,4 @@ __pycache__/
 *.egg
 *.log
 /data/*
+/results/*
